requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
entrypoints==0.2.3
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
idna==2.0
imutils==0.4.3
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.2
iso-639==0.4.5
iso3166==0.8
jedi==0.11.0
Jinja2==2.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
language-selector==0.1
livestreamer==1.12.2
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
pkg-resources==0.0.0
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==3.4.2
ptyprocess==0.5
pyasn1==0.1.9
pycryptodome==3.4.7
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
python-xapp==1.0.0
pytz==2017.2
pyxdg==0.25
pyzmq==16.0.2
qtconsole==4.3.1
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
setproctitle==1.1.8
simplegeneric==0.8.1
six==1.11.0
streamlink==0.8.1
system-service==0.3
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.12.2
widgetsnbextension==3.0.4
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
